+++ Cargo.toml
@@ -112,7 +112,9 @@ zip = { version = "0.6.6", default-features = false }
 
 pixi_config = { path = "crates/pixi_config" }
 pixi_consts = { path = "crates/pixi_consts" }
+pixi_default_sys = { path = "crates/pixi_default_sys" }
 pixi_manifest = { path = "crates/pixi_manifest" }
+pypi_modifiers = { path = "crates/pypi_modifiers" }
 
 [package]
 authors.workspace = true
@@ -214,7 +216,9 @@ rattler_solve = { workspace = true, features = ["resolvo", "serde"] }
 
 pixi_config = { workspace = true }
 pixi_consts = { workspace = true }
+pixi_default_sys = { workspace = true }
 pixi_manifest = { workspace = true }
+pypi_modifiers = { workspace = true }
 rattler_virtual_packages = { workspace = true }
 regex = { workspace = true }
 reqwest = { workspace = true, features = [
