+++ crates/pypi_modifiers/src/pypi_tags.rs
@@ -1,5 +1,5 @@
-use crate::project::virtual_packages::{default_glibc_version, default_mac_os_version};
 use miette::{Context, IntoDiagnostic};
+use pixi_default_sys::{default_glibc_version, default_mac_os_version};
 use pixi_manifest::{LibCSystemRequirement, SystemRequirements};
 use platform_tags::Os;
 use platform_tags::Tags;
