+++ src/project/virtual_packages.rs
@@ -3,6 +3,7 @@ use crate::project::errors::UnsupportedPlatformError;
 use crate::project::Environment;
 use itertools::Itertools;
 use miette::Diagnostic;
+use pixi_default_sys::{default_glibc_version, default_linux_version, default_mac_os_version};
 use pixi_manifest::{LibCSystemRequirement, SystemRequirements};
 use rattler_conda_types::{GenericVirtualPackage, Platform, Version};
 use rattler_virtual_packages::{
@@ -10,29 +11,6 @@ use rattler_virtual_packages::{
 };
 use std::collections::HashMap;
 use thiserror::Error;
-
-/// The default GLIBC version to use. This is used when no system requirements are specified.
-pub fn default_glibc_version() -> Version {
-    "2.28".parse().unwrap()
-}
-
-/// The default linux version to use. This is used when no system requirements are specified.
-pub fn default_linux_version() -> Version {
-    "5.10".parse().unwrap()
-}
-
-/// Returns the default Mac OS version for the specified platform. The platform must refer to a
-/// MacOS platform.
-pub fn default_mac_os_version(platform: Platform) -> Version {
-    match platform {
-        Platform::OsxArm64 => "13.0".parse().unwrap(),
-        Platform::Osx64 => "13.0".parse().unwrap(),
-        _ => panic!(
-            "default_mac_os_version() called with non-osx platform: {}",
-            platform
-        ),
-    }
-}
 
 /// Returns a reasonable modern set of virtual packages that should be safe enough to assume.
 /// At the time of writing, this is in sync with the conda-lock set of minimal virtual packages.
